casting total to float

+++ whitebox/utils.py
@@ -96,7 +96,7 @@ def create_insights(
     msedf = pd.DataFrame({'groupByValue': group.name,
                           'groupByVarName': group_var,
                           error_type: error_dict[error_type],
-                          'Total': int(group.shape[0])}, index=[0])
+                          'Total': float(group.shape[0])}, index=[0])
     return msedf
 
 def to_json(
